[ch57666] Add Haddock (#7)
