+++ src/LaunchDarkly/Server/Evaluate.hs
@@ -20,7 +20,7 @@ import GHC.Natural                         (Natural, naturalToInt)
 import Data.Word                           (Word8)
 import Data.ByteString                     (ByteString)
 
-import LaunchDarkly.Server.Client          (Client)
+import LaunchDarkly.Server.Client.Internal (Client)
 import LaunchDarkly.Server.User.Internal   (User, valueOf)
 import LaunchDarkly.Server.Features        (Flag, Segment, Prerequisite, SegmentRule, Clause, VariationOrRollout, Rule)
 import LaunchDarkly.Server.Store           (LaunchDarklyStoreRead, getFlag, getSegment)
